table veiw scroll bug fix
